chore: more tests
