--- Assets/PlayroomKit/ServiceHandler.cs
@@ -1,7 +1,0 @@
-namespace Playroom
-{
-    // public class ServiceHandler
-    // {
-    //     TODO: When partial class is refactored we use this to check which service to run based on current platform
-    // }
-}
