--- Assets/PlayroomKit/ServiceHandler.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: fab86cd0b40640ef80d9b6242746fa30
-timeCreated: 1729857326
