+++ Assets/PlayroomKit/modules/MockMode/BrowserMode/BrowserMockPlayerService.cs
@@ -5,7 +5,6 @@ using UnityEngine;
 namespace Playroom
 {
 #if UNITY_EDITOR
-
     public class BrowserMockPlayerService : PlayroomKit.Player.IPlayerBase
     {
         private readonly UnityBrowserBridge _ubb;
@@ -121,7 +120,6 @@ namespace Playroom
 
             _ubb.CallJs("WaitForPlayerState", null, null, true, _id, stateKey, callbackKey);
         }
-
 
         #region UTILS
